+++ packages/core-js-compat/src/modules-by-versions.js
@@ -73,4 +73,10 @@ module.exports = {
     'esnext.typed-array.at',
     'esnext.typed-array.filter-out',
   ],
+  3.9: [
+    'esnext.array.find-last',
+    'esnext.array.find-last-index',
+    'esnext.typed-array.find-last',
+    'esnext.typed-array.find-last-index',
+  ],
 };