+++ packages/core-js/features/typed-array/index.js
@@ -1,5 +1,7 @@
 var parent = require('../../es/typed-array');
 require('../../modules/esnext.typed-array.at');
 require('../../modules/esnext.typed-array.filter-out');
+require('../../modules/esnext.typed-array.find-last');
+require('../../modules/esnext.typed-array.find-last-index');
 
 module.exports = parent;